+++ Plugin/clSTCHelper.cpp
@@ -69,7 +69,7 @@ void clSTCHelper::UpdateScrollbarWidth(wxStyledTextCtrl* ctrl, size_t char_width
 
     wxString text;
     for (int i = startLine; i <= endLine; i++) {
-        int visibleLine = (int)ctrl->DocLineFromVisible(i); // get actual visible line, folding may offset lines
+        int visibleLine = ctrl->DocLineFromVisible(i); // get actual visible line, folding may offset lines
         wxString line_text = ctrl->GetLine(visibleLine);
         text = line_text.length() > text.length() ? line_text : text;
     }
